CDAP-15821 Whole-file-ingest plugin validation
