--- src/main/java/co/cask/hydrator/plugin/common/ReferencePluginConfig.java
@@ -1,36 +1,0 @@
-/*
- * Copyright © 2016 Cask Data, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-
-package co.cask.hydrator.plugin.common;
-
-import co.cask.cdap.api.annotation.Description;
-import co.cask.cdap.api.annotation.Name;
-import co.cask.cdap.api.plugin.PluginConfig;
-
-/**
- * {@link PluginConfig} that contains a referenceName property that is used to create an external Dataset for
- * metadata, lineage purposes.
- */
-public class ReferencePluginConfig extends PluginConfig {
-
-  @Name(Constants.Reference.REFERENCE_NAME)
-  @Description(Constants.Reference.REFERENCE_NAME_DESCRIPTION)
-  public String referenceName;
-
-  public ReferencePluginConfig(String referenceName) {
-    this.referenceName = referenceName;
-  }
-}
